+++ src/main/java/com/gameminers/mav/personality/poly/TrianglePersonality.java
@@ -15,9 +15,6 @@
  */
 package com.gameminers.mav.personality.poly;
 
-import org.lwjgl.opengl.GL11;
-
-import com.gameminers.mav.personality.Personality;
 import com.gameminers.mav.render.Rendering;
 
 public class TrianglePersonality extends PolygonPersonality {
@@ -30,7 +27,7 @@ public class TrianglePersonality extends PolygonPersonality {
 		float radius = diameter/2f;
 		float[] bg = Rendering.getColor(0.3f);
 		float[] fg = Rendering.getColor(0.8f);
-		Rendering.drawTriangle(0, 0, radius*(0.85f+(float)((pulse)*0.15f)), fg[0], fg[1], fg[2], 0.5f, 0);
+		Rendering.drawTriangle(0, 0, radius*(0.85f+(pulse)*0.15f), fg[0], fg[1], fg[2], 0.5f, 0);
 		Rendering.drawTriangle(0, 0, radius*0.8f, fg[0], fg[1], fg[2], 1, 0.5f);
 		Rendering.drawTriangle(0, 0, radius*0.7f, bg[0], bg[1], bg[2], 1, 1f);
 	}
